Reapply "Merge branch 'master' into bvrooman/feat/generic_key_size_sparse_merkle_tree"
This reverts commit 7b5f58c0e8f461659950915fe5e0af10ac18d1ec.

+++ fuel-tx/src/transaction/types/input/message.rs
@@ -35,7 +35,7 @@ pub trait MessageSpecification: private::Seal {
     type Predicate: AsField<Vec<u8>>;
     type PredicateData: AsField<Vec<u8>>;
     type PredicateGasUsed: AsField<Word>;
-    type Witness: AsField<u8>;
+    type Witness: AsField<u16>;
 }
 
 pub mod specifications {
@@ -72,7 +72,7 @@ pub mod specifications {
         type Predicate = Empty<Vec<u8>>;
         type PredicateData = Empty<Vec<u8>>;
         type PredicateGasUsed = Empty<Word>;
-        type Witness = u8;
+        type Witness = u16;
     }
 
     impl MessageSpecification for MessageData<Predicate> {
@@ -80,7 +80,7 @@ pub mod specifications {
         type Predicate = Vec<u8>;
         type PredicateData = Vec<u8>;
         type PredicateGasUsed = Word;
-        type Witness = Empty<u8>;
+        type Witness = Empty<u16>;
     }
 
     /// The spendable message acts as a standard coin.
@@ -93,7 +93,7 @@ pub mod specifications {
         type Predicate = Empty<Vec<u8>>;
         type PredicateData = Empty<Vec<u8>>;
         type PredicateGasUsed = Empty<Word>;
-        type Witness = u8;
+        type Witness = u16;
     }
 
     impl MessageSpecification for MessageCoin<Predicate> {
@@ -101,7 +101,7 @@ pub mod specifications {
         type Predicate = Vec<u8>;
         type PredicateData = Vec<u8>;
         type PredicateGasUsed = Word;
-        type Witness = Empty<u8>;
+        type Witness = Empty<u16>;
     }
 
     /// The type is used to represent the full message. It is used during the
@@ -119,7 +119,7 @@ pub mod specifications {
         type Predicate = Vec<u8>;
         type PredicateData = Vec<u8>;
         type PredicateGasUsed = Word;
-        type Witness = u8;
+        type Witness = u16;
     }
 }
